move module tool button implementation into ui/module_tool_button.py
