+++ ballontranslator/ui/mainwindowbars.py
@@ -1,25 +1,44 @@
-import json
-import os
 import os.path as osp
 from typing import List, Union
 
-from qtpy.QtWidgets import QMainWindow, QHBoxLayout, QVBoxLayout, QFileDialog, QLabel, QSizePolicy, QToolBar, QMenu, QSpacerItem, QPushButton, QCheckBox, QToolButton, QWidgetAction
-from qtpy.QtCore import Qt, Signal, QPoint, QEvent, QSize, QRectF
-from qtpy.QtGui import QMouseEvent, QKeySequence, QActionGroup, QIcon, QPainterPath, QRegion
+from qtpy.QtWidgets import (
+    QCheckBox,
+    QFileDialog,
+    QHBoxLayout,
+    QLabel,
+    QMainWindow,
+    QMenu,
+    QPushButton,
+    QSizePolicy,
+    QSpacerItem,
+    QToolBar,
+    QToolButton,
+    QVBoxLayout,
+)
+from qtpy.QtCore import Qt, Signal, QPoint, QEvent
+from qtpy.QtGui import QMouseEvent, QKeySequence, QActionGroup
 
-from .custom_widget import Widget, PaintQSlider, SmallComboBox
-from .misc import themed_icon_path
-from ballontranslator.utils.shared import TITLEBAR_HEIGHT, WINDOW_BORDER_WIDTH, BOTTOMBAR_HEIGHT, LEFTBAR_WIDTH, LEFTBTN_WIDTH
+from .custom_widget import Widget, PaintQSlider
+from .module_tool_button import (
+    InpaintSelectionWidget,
+    ModuleSelectionToolButtonWidget,
+    OCRSelectionWidget,
+    TranslatorSelectionWidget,
+)
+from ballontranslator.utils.shared import (
+    BOTTOMBAR_HEIGHT,
+    LEFTBAR_WIDTH,
+    LEFTBTN_WIDTH,
+    TITLEBAR_HEIGHT,
+    WINDOW_BORDER_WIDTH,
+)
 from .framelesswindow import FramelessMoveResize
 from ballontranslator.utils.config import pcfg
-from ballontranslator.utils.llm_profiles import LLM_INPAINT_KEY, LLM_OCR_KEY, LLM_TRANSLATOR_KEY, profile_by_id
 from ballontranslator.utils import shared
 if shared.FLAG_QT6:
     from qtpy.QtGui import QAction
 else:
     from qtpy.QtWidgets import QAction
-
-BOTTOM_BAR_ACCENT_COLOR = '#5e98f7'
 
 class ShowPageListChecker(QCheckBox):
     ...
@@ -537,784 +556,6 @@ class TitleBar(Widget):
             title += ' - '  + self.save_state
         self.titleLabel.setText(title)
 
-
-class SmallConfigPutton(QPushButton):
-    pass
-
-
-def _set_bottom_aux_button_visible(button: QPushButton, visible: bool):
-    if visible == (not button.isHidden()):
-        return
-    button.setVisible(visible)
-    parent = button.parentWidget()
-    if parent is not None:
-        layout = parent.layout()
-        if layout is not None:
-            layout.invalidate()
-        parent.updateGeometry()
-
-
-def cfg_icon() -> QIcon:
-    return QIcon(themed_icon_path('leftbar_config_activate.svg'))
-
-
-def _theme_value(key: str, fallback: str) -> str:
-    theme = 'eva-dark' if pcfg.darkmode else 'eva-light'
-    try:
-        with open(shared.THEME_PATH, 'r', encoding='utf8') as f:
-            theme_dict = json.loads(f.read())
-        return theme_dict.get(theme, {}).get(key, fallback)
-    except Exception:
-        return fallback
-
-
-def _theme_foreground_hex() -> str:
-    fallback = '#8e99b1' if pcfg.darkmode else '#5d5d5f'
-    return _theme_value('@qwidgetForegroundColor', fallback)
-
-
-def _theme_menu_background_hex() -> str:
-    fallback = '#21252b' if pcfg.darkmode else '#e1e4eb'
-    return _theme_value('@emptyContentBackgroundColor', fallback)
-
-
-def _blend_hex(color: str, target: str, amount: float) -> str:
-    def rgb(value: str):
-        value = value.lstrip('#')
-        return [int(value[i:i + 2], 16) for i in (0, 2, 4)]
-
-    src = rgb(color)
-    dst = rgb(target)
-    mixed = [round(src[i] + (dst[i] - src[i]) * amount) for i in range(3)]
-    return '#{:02x}{:02x}{:02x}'.format(*mixed)
-
-
-def _section_label_color() -> str:
-    foreground = _theme_foreground_hex()
-    return _blend_hex(foreground, '#000000' if pcfg.darkmode else '#ffffff', 0.28)
-
-
-def _bottom_tool_icon_path(icon_filename: str) -> str:
-    icon_path = themed_icon_path(icon_filename)
-    if not pcfg.darkmode or not icon_filename.lower().endswith('.svg'):
-        return icon_path
-
-    color = _theme_foreground_hex()
-    cache_dir = osp.join(shared.cache_dir, 'icons', 'bottom_bar', 'eva-dark')
-    os.makedirs(cache_dir, exist_ok=True)
-    cache_path = osp.join(cache_dir, icon_filename)
-    with open(icon_path, 'r', encoding='utf-8') as f:
-        svg = f.read()
-    for fill in ('#697187', '#b3b6bf', '#96a4cd', '#697186'):
-        svg = svg.replace(fill, color)
-    with open(cache_path, 'w', encoding='utf-8') as f:
-        f.write(svg)
-    return cache_path
-
-
-def _set_bottom_tool_button_icon(tool_btn: QToolButton, icon_filename: str):
-    tool_btn.setIcon(QIcon(_bottom_tool_icon_path(icon_filename)))
-    tool_btn.setIconSize(QSize(18, 18))
-    style_enum = getattr(Qt, 'ToolButtonStyle', Qt)
-    tool_btn.setToolButtonStyle(style_enum.ToolButtonTextBesideIcon)
-
-
-def _instant_popup_mode():
-    popup_enum = getattr(QToolButton, 'ToolButtonPopupMode', QToolButton)
-    return popup_enum.InstantPopup
-
-
-def _bottom_tool_button_text(name: str) -> str:
-    return '  ' + name
-
-
-def _simplify_llm_model_name(model: str) -> str:
-    parts = [part.strip() for part in str(model or '').split('/') if part.strip()]
-    return parts[-1] if parts else ''
-
-
-
-class BottomBarMenu(QMenu):
-
-    def showEvent(self, event) -> None:
-        super().showEvent(event)
-        path = QPainterPath()
-        path.addRoundedRect(QRectF(self.rect()), 8, 8)
-        self.setMask(QRegion(path.toFillPolygon().toPolygon()))
-
-
-def _style_bottom_menu(menu: QMenu) -> QMenu:
-    menu.setObjectName('BottomBarModuleMenu')
-    attr_enum = getattr(Qt, 'WidgetAttribute', Qt)
-    menu.setAttribute(attr_enum.WA_TranslucentBackground, True)
-    return menu
-
-
-def _bottom_menu(parent: QToolButton) -> QMenu:
-    return _style_bottom_menu(BottomBarMenu(parent))
-
-
-def _bottom_submenu(title: str, parent: QMenu) -> QMenu:
-    menu = _style_bottom_menu(BottomBarMenu(parent))
-    menu.setTitle(title)
-    return menu
-
-
-def _add_bottom_menu_section(menu: QMenu, text: str, accent: bool = False):
-    label = QLabel(text, menu)
-    label.setObjectName('BottomBarMenuSectionLabel')
-    color = BOTTOM_BAR_ACCENT_COLOR if accent else _section_label_color()
-    label.setStyleSheet(
-        'QLabel#BottomBarMenuSectionLabel {{ '
-        'color: {}; background-color: {}; '
-        '}}'.format(color, _theme_menu_background_hex())
-    )
-    action = QWidgetAction(menu)
-    action.setDefaultWidget(label)
-    menu.addAction(action)
-
-
-def _checked_action_text(text: str, checked: bool) -> str:
-    return text + ('\t\u2713' if checked else '')
-
-
-def _add_bottom_menu_action(menu: QMenu, text: str, checked: bool, callback):
-    action = QAction(_checked_action_text(text, checked), menu)
-    action.triggered.connect(callback)
-    menu.addAction(action)
-    return action
-
-
-def _add_bottom_submenu(parent: QMenu, submenu: QMenu, text: str, checked: bool):
-    parent.addMenu(submenu)
-    submenu.menuAction().setText(_checked_action_text(text, checked))
-    return submenu
-
-
-class ModuleSelectionToolButtonWidget(Widget):
-
-    cfg_clicked = Signal()
-
-    def __init__(self, fallback_name: str, icon_filename: str, *args, **kwargs) -> None:
-        super().__init__(*args, **kwargs)
-        self.fallback_name = fallback_name
-        self.icon_filename = icon_filename
-        self.selector = SmallComboBox()
-        self.selector.setVisible(False)
-        self.selector.currentTextChanged.connect(self.updateButtonText)
-
-        self.tool_btn = QToolButton(self)
-        self.tool_btn.setObjectName('BottomBarModuleToolButton')
-        self.tool_btn.setToolTip(fallback_name)
-        self.tool_btn.setPopupMode(_instant_popup_mode())
-        _set_bottom_tool_button_icon(self.tool_btn, icon_filename)
-        self.tool_btn.setText(fallback_name)
-        self.menu = _bottom_menu(self.tool_btn)
-        self.tool_btn.setMenu(self.menu)
-        self.menu.aboutToShow.connect(self.rebuildMenu)
-
-        self.cfg_btn = SmallConfigPutton()
-        self.cfg_btn.clicked.connect(self.cfg_clicked)
-        self.cfg_btn.setVisible(False)
-
-        layout = QHBoxLayout(self)
-        layout.setContentsMargins(0, 0, 0, 0)
-        layout.setSpacing(1)
-        layout.addWidget(self.tool_btn)
-        layout.addWidget(self.cfg_btn)
-        self.updateButtonText()
-
-    def enterEvent(self, event: QEvent) -> None:
-        _set_bottom_aux_button_visible(self.cfg_btn, True)
-        self.cfg_btn.setIcon(cfg_icon())
-        return super().enterEvent(event)
-
-    def leaveEvent(self, event: QEvent) -> None:
-        self.cfg_btn.setIcon(QIcon())
-        _set_bottom_aux_button_visible(self.cfg_btn, False)
-        return super().leaveEvent(event)
-
-    def changeEvent(self, event: QEvent) -> None:
-        if event.type() in (QEvent.Type.StyleChange, QEvent.Type.PaletteChange):
-            _set_bottom_tool_button_icon(self.tool_btn, self.icon_filename)
-        return super().changeEvent(event)
-
-    def blockSignals(self, block: bool):
-        self.selector.blockSignals(block)
-        super().blockSignals(block)
-
-    def setSelectedValue(self, value: str, block_signals=True):
-        if block_signals:
-            self.blockSignals(True)
-        self.selector.setCurrentText(value)
-        if block_signals:
-            self.blockSignals(False)
-        self.updateButtonText()
-
-    def rebuildMenu(self):
-        self.menu.clear()
-        current_module = self.selector.currentText()
-        for i in range(self.selector.count()):
-            module = self.selector.itemText(i)
-            _add_bottom_menu_action(
-                self.menu,
-                module,
-                module == current_module,
-                lambda checked=False, value=module: self.selector.setCurrentText(value),
-            )
-
-    def updateButtonText(self, *args):
-        name = self.selector.currentText() or self.fallback_name
-        self.tool_btn.setText(_bottom_tool_button_text(name))
-
-
-class TranslatorSelectionWidget(Widget):
-
-    cfg_clicked = Signal()
-    edit_clicked = Signal(str)
-    llm_profile_changed = Signal(str)
-
-    def __init__(self) -> None:
-        super().__init__()
-        self.selector = SmallComboBox()
-        self.src_selector = SmallComboBox()
-        self.tgt_selector = SmallComboBox()
-        self.selector.setVisible(False)
-        self.src_selector.setVisible(False)
-        self.tgt_selector.setVisible(False)
-        self.tool_btn = QToolButton(self)
-        self.tool_btn.setObjectName('BottomBarModuleToolButton')
-        self.tool_btn.setToolTip(self.tr('Translator'))
-        self.tool_btn.setPopupMode(_instant_popup_mode())
-        self.tool_btn.setText(self.tr('Translator'))
-        self.icon_filename = 'bottombar_translate_activate.svg'
-        _set_bottom_tool_button_icon(self.tool_btn, self.icon_filename)
-        self.menu = _bottom_menu(self.tool_btn)
-        self.tool_btn.setMenu(self.menu)
-        self.menu.aboutToShow.connect(self.rebuildMenu)
-        self.edit_btn = SmallConfigPutton()
-        self.edit_btn.clicked.connect(self.onEditClicked)
-        self.cfg_btn = SmallConfigPutton()
-        self.cfg_btn.clicked.connect(self.cfg_clicked)
-        self.edit_btn.setVisible(False)
-        self.cfg_btn.setVisible(False)
-
-        layout = QHBoxLayout(self)
-        layout.setContentsMargins(0, 0, 0, 0)
-        layout.addWidget(self.tool_btn)
-        layout.addWidget(self.edit_btn)
-        layout.addWidget(self.cfg_btn)
-        layout.setSpacing(1)
-        self.updateButtonText()
-
-    def enterEvent(self, event: QEvent) -> None:
-        show_edit = self.selector.currentText() == LLM_TRANSLATOR_KEY
-        _set_bottom_aux_button_visible(self.edit_btn, show_edit)
-        if show_edit:
-            self.edit_btn.setIcon(QIcon(themed_icon_path('edit.svg')))
-        _set_bottom_aux_button_visible(self.cfg_btn, True)
-        self.cfg_btn.setIcon(cfg_icon())
-        return super().enterEvent(event)
-
-    def leaveEvent(self, event: QEvent) -> None:
-        self.edit_btn.setIcon(QIcon())
-        self.cfg_btn.setIcon(QIcon())
-        _set_bottom_aux_button_visible(self.edit_btn, False)
-        _set_bottom_aux_button_visible(self.cfg_btn, False)
-        return super().leaveEvent(event)
-
-    def changeEvent(self, event: QEvent) -> None:
-        if event.type() in (QEvent.Type.StyleChange, QEvent.Type.PaletteChange):
-            _set_bottom_tool_button_icon(self.tool_btn, self.icon_filename)
-        return super().changeEvent(event)
-    
-    def blockSignals(self, block: bool):
-        self.src_selector.blockSignals(block)
-        self.tgt_selector.blockSignals(block)
-        self.selector.blockSignals(block)
-        super().blockSignals(block)
-
-    def _section(self, text: str, accent: bool = False):
-        _add_bottom_menu_section(self.menu, text, accent=accent)
-
-    def rebuildMenu(self):
-        self.menu.clear()
-        current_translator = self.selector.currentText()
-        self._section(self.tr('Translator'))
-        for i in range(self.selector.count()):
-            translator = self.selector.itemText(i)
-            if translator == LLM_TRANSLATOR_KEY:
-                continue
-            _add_bottom_menu_action(
-                self.menu,
-                translator,
-                current_translator == translator,
-                lambda checked=False, value=translator: self.selector.setCurrentText(value),
-            )
-
-        self._section(self.tr('LLM'), accent=current_translator == LLM_TRANSLATOR_KEY)
-        added = False
-        for profile in pcfg.module.llm_profiles:
-            if not profile.support_text:
-                continue
-            added = True
-            profile_id = profile.id
-            profile_menu = _bottom_submenu(profile.name or profile_id, self.menu)
-            _add_bottom_submenu(
-                self.menu,
-                profile_menu,
-                profile.name or profile_id,
-                current_translator == LLM_TRANSLATOR_KEY and pcfg.module.translator_llm_id == profile_id,
-            )
-            self._buildProfileMenu(profile_menu, profile)
-        if not added:
-            action = QAction(self.tr('No text profiles'), self.menu)
-            action.setEnabled(False)
-            self.menu.addAction(action)
-
-        self._section(self.tr('Language'))
-        source_menu = _bottom_submenu(
-            self.tr('Source - {language}').format(language=self.src_selector.currentText()),
-            self.menu,
-        )
-        self.menu.addMenu(source_menu)
-        for i in range(self.src_selector.count()):
-            lang = self.src_selector.itemText(i)
-            _add_bottom_menu_action(
-                source_menu,
-                lang,
-                lang == self.src_selector.currentText(),
-                lambda checked=False, value=lang: self.src_selector.setCurrentText(value),
-            )
-
-        target_menu = _bottom_submenu(
-            self.tr('Target - {language}').format(language=self.tgt_selector.currentText()),
-            self.menu,
-        )
-        self.menu.addMenu(target_menu)
-        for i in range(self.tgt_selector.count()):
-            lang = self.tgt_selector.itemText(i)
-            _add_bottom_menu_action(
-                target_menu,
-                lang,
-                lang == self.tgt_selector.currentText(),
-                lambda checked=False, value=lang: self.tgt_selector.setCurrentText(value),
-            )
-
-    def selectLLMProfile(self, profile_id: str):
-        pcfg.module.translator_llm_id = profile_id
-        if self.selector.currentText() != LLM_TRANSLATOR_KEY:
-            self.selector.setCurrentText(LLM_TRANSLATOR_KEY)
-        self.llm_profile_changed.emit(profile_id)
-        self.updateButtonText()
-
-    def selectLLMProfileSetting(self, profile_id: str, key: str, value: str):
-        profile = profile_by_id(pcfg.module.llm_profiles, profile_id)
-        if profile is not None:
-            setattr(profile, key, value)
-            if key == 'model':
-                options = profile.model_options
-                if value and value not in options:
-                    options.insert(0, value)
-        self.selectLLMProfile(profile_id)
-
-    def _buildProfileMenu(self, menu: QMenu, profile: dict):
-        profile_id = profile.id
-        selected_profile = (
-            self.selector.currentText() == LLM_TRANSLATOR_KEY
-            and pcfg.module.translator_llm_id == profile_id
-        )
-
-        _add_bottom_menu_section(menu, self.tr('Thinking Level'))
-        thinking_options = [str(option) for option in profile.thinking_level_options if str(option)]
-        current_thinking = str(profile.thinking_level or 'None')
-        for thinking_level in thinking_options:
-            _add_bottom_menu_action(
-                menu,
-                thinking_level,
-                selected_profile and thinking_level == current_thinking,
-                lambda checked=False, pid=profile_id, value=thinking_level: self.selectLLMProfileSetting(pid, 'thinking_level', value),
-            )
-
-        _add_bottom_menu_section(menu, self.tr('Model'))
-        model_options = [str(option) for option in profile.model_options if str(option)]
-        current_model = str(profile.model or '')
-        for model in model_options:
-            _add_bottom_menu_action(
-                menu,
-                model,
-                selected_profile and model == current_model,
-                lambda checked=False, pid=profile_id, value=model: self.selectLLMProfileSetting(pid, 'model', value),
-            )
-
-    def onEditClicked(self):
-        if self.selector.currentText() == LLM_TRANSLATOR_KEY:
-            self.edit_clicked.emit(pcfg.module.translator_llm_id)
-
-    def updateButtonText(self):
-        name = self.selector.currentText()
-        is_llm = name == LLM_TRANSLATOR_KEY
-        if name == LLM_TRANSLATOR_KEY:
-            profile = profile_by_id(pcfg.module.llm_profiles, pcfg.module.translator_llm_id)
-            if profile is not None:
-                model_options = [str(option) for option in profile.model_options if str(option)]
-                model = str(profile.model or '').strip()
-                thinking_level = str(profile.thinking_level or 'None').strip()
-                if model_options and model:
-                    name = _simplify_llm_model_name(model)
-                    if thinking_level and thinking_level != 'None':
-                        name = self.tr('{model} {thinking_level}').format(model=name, thinking_level=thinking_level)
-                else:
-                    name = profile.name or name
-        if not name:
-            name = self.tr('Translator')
-        self.tool_btn.setText(_bottom_tool_button_text(name))
-        _set_bottom_aux_button_visible(self.edit_btn, is_llm and self.underMouse())
-    
-    def setTranslatorMetadata(self, name: str, supported_src_list, supported_tgt_list, lang_source: str, lang_target: str):
-        # Metadata can come from ModuleSpec before the translator is imported.
-        self.blockSignals(True)
-        self.src_selector.clear()
-        self.tgt_selector.clear()
-        self.src_selector.addItems(supported_src_list)
-        self.tgt_selector.addItems(supported_tgt_list)
-        self.selector.setCurrentText(name)
-        self.src_selector.setCurrentText(lang_source)
-        self.tgt_selector.setCurrentText(lang_target)
-        self.blockSignals(False)
-        self.updateButtonText()
-
-
-class OCRSelectionWidget(Widget):
-
-    cfg_clicked = Signal()
-    edit_clicked = Signal(str)
-    llm_profile_changed = Signal(str)
-
-    def __init__(self) -> None:
-        super().__init__()
-        self.selector = SmallComboBox()
-        self.selector.setVisible(False)
-        self.selector.currentTextChanged.connect(self.updateButtonText)
-        self.tool_btn = QToolButton(self)
-        self.tool_btn.setObjectName('BottomBarModuleToolButton')
-        self.tool_btn.setToolTip(self.tr('OCR'))
-        self.tool_btn.setPopupMode(_instant_popup_mode())
-        self.icon_filename = 'small_ocr.svg'
-        _set_bottom_tool_button_icon(self.tool_btn, self.icon_filename)
-        self.tool_btn.setText(self.tr('OCR'))
-        self.menu = _bottom_menu(self.tool_btn)
-        self.tool_btn.setMenu(self.menu)
-        self.menu.aboutToShow.connect(self.rebuildMenu)
-        self.edit_btn = SmallConfigPutton()
-        self.edit_btn.clicked.connect(self.onEditClicked)
-        self.cfg_btn = SmallConfigPutton()
-        self.cfg_btn.clicked.connect(self.cfg_clicked)
-        self.edit_btn.setVisible(False)
-        self.cfg_btn.setVisible(False)
-
-        layout = QHBoxLayout(self)
-        layout.setContentsMargins(0, 0, 0, 0)
-        layout.setSpacing(1)
-        layout.addWidget(self.tool_btn)
-        layout.addWidget(self.edit_btn)
-        layout.addWidget(self.cfg_btn)
-        self.updateButtonText()
-
-    def enterEvent(self, event: QEvent) -> None:
-        show_edit = self.selector.currentText() == LLM_OCR_KEY
-        _set_bottom_aux_button_visible(self.edit_btn, show_edit)
-        if show_edit:
-            self.edit_btn.setIcon(QIcon(themed_icon_path('edit.svg')))
-        _set_bottom_aux_button_visible(self.cfg_btn, True)
-        self.cfg_btn.setIcon(cfg_icon())
-        return super().enterEvent(event)
-
-    def leaveEvent(self, event: QEvent) -> None:
-        self.edit_btn.setIcon(QIcon())
-        self.cfg_btn.setIcon(QIcon())
-        _set_bottom_aux_button_visible(self.edit_btn, False)
-        _set_bottom_aux_button_visible(self.cfg_btn, False)
-        return super().leaveEvent(event)
-
-    def changeEvent(self, event: QEvent) -> None:
-        if event.type() in (QEvent.Type.StyleChange, QEvent.Type.PaletteChange):
-            _set_bottom_tool_button_icon(self.tool_btn, self.icon_filename)
-        return super().changeEvent(event)
-
-    def blockSignals(self, block: bool):
-        self.selector.blockSignals(block)
-        super().blockSignals(block)
-
-    def setSelectedValue(self, value: str, block_signals=True):
-        if block_signals:
-            self.blockSignals(True)
-        self.selector.setCurrentText(value)
-        if block_signals:
-            self.blockSignals(False)
-        self.updateButtonText()
-
-    def _section(self, text: str, accent: bool = False):
-        _add_bottom_menu_section(self.menu, text, accent=accent)
-
-    def rebuildMenu(self):
-        self.menu.clear()
-        current_ocr = self.selector.currentText()
-        self._section(self.tr('OCR'))
-        for i in range(self.selector.count()):
-            ocr = self.selector.itemText(i)
-            if ocr == LLM_OCR_KEY:
-                continue
-            _add_bottom_menu_action(
-                self.menu,
-                ocr,
-                current_ocr == ocr,
-                lambda checked=False, value=ocr: self.selector.setCurrentText(value),
-            )
-
-        self._section(self.tr('LLM'), accent=current_ocr == LLM_OCR_KEY)
-        added = False
-        for profile in pcfg.module.llm_profiles:
-            if not profile.support_vision:
-                continue
-            added = True
-            profile_id = profile.id
-            profile_menu = _bottom_submenu(profile.name or profile_id, self.menu)
-            _add_bottom_submenu(
-                self.menu,
-                profile_menu,
-                profile.name or profile_id,
-                current_ocr == LLM_OCR_KEY and pcfg.module.ocr_llm_id == profile_id,
-            )
-            self._buildProfileMenu(profile_menu, profile)
-        if not added:
-            action = QAction(self.tr('No vision profiles'), self.menu)
-            action.setEnabled(False)
-            self.menu.addAction(action)
-
-    def selectLLMProfile(self, profile_id: str):
-        pcfg.module.ocr_llm_id = profile_id
-        if self.selector.currentText() != LLM_OCR_KEY:
-            self.selector.setCurrentText(LLM_OCR_KEY)
-        self.llm_profile_changed.emit(profile_id)
-        self.updateButtonText()
-
-    def selectLLMProfileSetting(self, profile_id: str, key: str, value: str):
-        profile = profile_by_id(pcfg.module.llm_profiles, profile_id)
-        if profile is not None:
-            setattr(profile, key, value)
-            if key == 'vision_model':
-                options = profile.vision_model_options
-                if value and value not in options:
-                    options.insert(0, value)
-        self.selectLLMProfile(profile_id)
-
-    def _buildProfileMenu(self, menu: QMenu, profile: dict):
-        profile_id = profile.id
-        selected_profile = (
-            self.selector.currentText() == LLM_OCR_KEY
-            and pcfg.module.ocr_llm_id == profile_id
-        )
-
-        _add_bottom_menu_section(menu, self.tr('Vision Model'))
-        model_options = [str(option) for option in profile.vision_model_options if str(option)]
-        current_model = str(profile.vision_model or '')
-        for model in model_options:
-            _add_bottom_menu_action(
-                menu,
-                model,
-                selected_profile and model == current_model,
-                lambda checked=False, pid=profile_id, value=model: self.selectLLMProfileSetting(pid, 'vision_model', value),
-            )
-
-        _add_bottom_menu_section(menu, self.tr('Vision Detail Level'))
-        current_detail = str(profile.vision_detail_level or 'None')
-        detail_options = [str(option) for option in profile.vision_detail_level_options if str(option)]
-        for detail_level in detail_options:
-            _add_bottom_menu_action(
-                menu,
-                detail_level,
-                selected_profile and detail_level == current_detail,
-                lambda checked=False, pid=profile_id, value=detail_level: self.selectLLMProfileSetting(pid, 'vision_detail_level', value),
-            )
-
-    def onEditClicked(self):
-        if self.selector.currentText() == LLM_OCR_KEY:
-            self.edit_clicked.emit(pcfg.module.ocr_llm_id)
-
-    def updateButtonText(self, *args):
-        name = self.selector.currentText()
-        is_llm = name == LLM_OCR_KEY
-        if is_llm:
-            profile = profile_by_id(pcfg.module.llm_profiles, pcfg.module.ocr_llm_id)
-            if profile is not None:
-                model = str(profile.vision_model or '').strip()
-                name = _simplify_llm_model_name(model) or profile.name or name
-        if not name:
-            name = self.tr('OCR')
-        self.tool_btn.setText(_bottom_tool_button_text(name))
-        _set_bottom_aux_button_visible(self.edit_btn, is_llm and self.underMouse())
-
-
-class InpaintSelectionWidget(Widget):
-
-    cfg_clicked = Signal()
-    edit_clicked = Signal(str)
-    llm_profile_changed = Signal(str)
-
-    def __init__(self) -> None:
-        super().__init__()
-        self.selector = SmallComboBox()
-        self.selector.setVisible(False)
-        self.selector.currentTextChanged.connect(self.updateButtonText)
-        self.tool_btn = QToolButton(self)
-        self.tool_btn.setObjectName('BottomBarModuleToolButton')
-        self.tool_btn.setToolTip(self.tr('Inpaint'))
-        self.tool_btn.setPopupMode(_instant_popup_mode())
-        self.icon_filename = 'drawingtools_inpaint.svg'
-        _set_bottom_tool_button_icon(self.tool_btn, self.icon_filename)
-        self.tool_btn.setText(self.tr('Inpaint'))
-        self.menu = _bottom_menu(self.tool_btn)
-        self.tool_btn.setMenu(self.menu)
-        self.menu.aboutToShow.connect(self.rebuildMenu)
-        self.edit_btn = SmallConfigPutton()
-        self.edit_btn.clicked.connect(self.onEditClicked)
-        self.cfg_btn = SmallConfigPutton()
-        self.cfg_btn.clicked.connect(self.cfg_clicked)
-        self.edit_btn.setVisible(False)
-        self.cfg_btn.setVisible(False)
-
-        layout = QHBoxLayout(self)
-        layout.setContentsMargins(0, 0, 0, 0)
-        layout.setSpacing(1)
-        layout.addWidget(self.tool_btn)
-        layout.addWidget(self.edit_btn)
-        layout.addWidget(self.cfg_btn)
-        self.updateButtonText()
-
-    def enterEvent(self, event: QEvent) -> None:
-        show_edit = self.selector.currentText() == LLM_INPAINT_KEY
-        _set_bottom_aux_button_visible(self.edit_btn, show_edit)
-        if show_edit:
-            self.edit_btn.setIcon(QIcon(themed_icon_path('edit.svg')))
-        _set_bottom_aux_button_visible(self.cfg_btn, True)
-        self.cfg_btn.setIcon(cfg_icon())
-        return super().enterEvent(event)
-
-    def leaveEvent(self, event: QEvent) -> None:
-        self.edit_btn.setIcon(QIcon())
-        self.cfg_btn.setIcon(QIcon())
-        _set_bottom_aux_button_visible(self.edit_btn, False)
-        _set_bottom_aux_button_visible(self.cfg_btn, False)
-        return super().leaveEvent(event)
-
-    def changeEvent(self, event: QEvent) -> None:
-        if event.type() in (QEvent.Type.StyleChange, QEvent.Type.PaletteChange):
-            _set_bottom_tool_button_icon(self.tool_btn, self.icon_filename)
-        return super().changeEvent(event)
-
-    def blockSignals(self, block: bool):
-        self.selector.blockSignals(block)
-        super().blockSignals(block)
-
-    def setSelectedValue(self, value: str, block_signals=True):
-        if block_signals:
-            self.blockSignals(True)
-        self.selector.setCurrentText(value)
-        if block_signals:
-            self.blockSignals(False)
-        self.updateButtonText()
-
-    def _section(self, text: str, accent: bool = False):
-        _add_bottom_menu_section(self.menu, text, accent=accent)
-
-    def rebuildMenu(self):
-        self.menu.clear()
-        current_inpaint = self.selector.currentText()
-        self._section(self.tr('Inpaint'))
-        for i in range(self.selector.count()):
-            inpainter = self.selector.itemText(i)
-            if inpainter == LLM_INPAINT_KEY:
-                continue
-            _add_bottom_menu_action(
-                self.menu,
-                inpainter,
-                current_inpaint == inpainter,
-                lambda checked=False, value=inpainter: self.selector.setCurrentText(value),
-            )
-
-        self._section(self.tr('LLM'), accent=current_inpaint == LLM_INPAINT_KEY)
-        added = False
-        for profile in pcfg.module.llm_profiles:
-            if not profile.support_image:
-                continue
-            added = True
-            profile_id = profile.id
-            profile_menu = _bottom_submenu(profile.name or profile_id, self.menu)
-            _add_bottom_submenu(
-                self.menu,
-                profile_menu,
-                profile.name or profile_id,
-                current_inpaint == LLM_INPAINT_KEY and pcfg.module.inpaint_llm_id == profile_id,
-            )
-            self._buildProfileMenu(profile_menu, profile)
-        if not added:
-            action = QAction(self.tr('No image profiles'), self.menu)
-            action.setEnabled(False)
-            self.menu.addAction(action)
-
-    def selectLLMProfile(self, profile_id: str):
-        pcfg.module.inpaint_llm_id = profile_id
-        pcfg.module.inpainter = LLM_INPAINT_KEY
-        if self.selector.currentText() != LLM_INPAINT_KEY:
-            self.selector.setCurrentText(LLM_INPAINT_KEY)
-        self.llm_profile_changed.emit(profile_id)
-        self.updateButtonText()
-
-    def selectLLMProfileSetting(self, profile_id: str, key: str, value: str):
-        profile = profile_by_id(pcfg.module.llm_profiles, profile_id)
-        if profile is not None:
-            setattr(profile, key, value)
-            if key == 'image_model':
-                options = profile.image_model_options
-                if value and value not in options:
-                    options.insert(0, value)
-        self.selectLLMProfile(profile_id)
-
-    def _buildProfileMenu(self, menu: QMenu, profile: dict):
-        profile_id = profile.id
-        selected_profile = (
-            self.selector.currentText() == LLM_INPAINT_KEY
-            and pcfg.module.inpaint_llm_id == profile_id
-        )
-
-        _add_bottom_menu_section(menu, self.tr('Image Model'))
-        model_options = [str(option) for option in profile.image_model_options if str(option)]
-        current_model = str(profile.image_model or '')
-        for model in model_options:
-            _add_bottom_menu_action(
-                menu,
-                model,
-                selected_profile and model == current_model,
-                lambda checked=False, pid=profile_id, value=model: self.selectLLMProfileSetting(pid, 'image_model', value),
-            )
-
-    def onEditClicked(self):
-        if self.selector.currentText() == LLM_INPAINT_KEY:
-            self.edit_clicked.emit(pcfg.module.inpaint_llm_id)
-
-    def updateButtonText(self, *args):
-        name = self.selector.currentText()
-        is_llm = name == LLM_INPAINT_KEY
-        if is_llm:
-            profile = profile_by_id(pcfg.module.llm_profiles, pcfg.module.inpaint_llm_id)
-            if profile is not None:
-                model = str(profile.image_model or '').strip()
-                name = _simplify_llm_model_name(model) or profile.name or name
-        if not name:
-            name = self.tr('Inpaint')
-        self.tool_btn.setText(_bottom_tool_button_text(name))
-        _set_bottom_aux_button_visible(self.edit_btn, is_llm and self.underMouse())
